refactor(dev): proxy websocket via native node upgrade, bridge only on bun/deno
Node proxies upgrades over the `http.Server` "upgrade" event + raw socket
(httpxy) directly — transparent and single-framed. The crossws WebSocket
bridge is only needed on Bun/Deno, which serve natively and expose no Node
upgrade socket. Inline the proxy helper into the dev server as an internal util.
